+++ src/controllers/index.js
@@ -1,83 +1,10 @@
-const {
-  Unary,
-  ServerStreaming,
-  BothWayStreaming,
-  getType
-} = require('../proto/EndpointTypes');
-
-const DEFAULT_STREAMING_DELAY = 1000;
-
-const unaryController = (endpoint, mappingResolver, dataFiles) => {
-  return {
-    endpointId: endpoint.getId(),
-    canHandle: (endpointId) => {
-      return endpoint.getId() === endpointId;
-    },
-    handle : async (callContext, callback) => {
-      const request = callContext.request;
-      const responseFile = mappingResolver.getResponseFile(endpoint.getId(), request);
-      const template = await dataFiles.get(responseFile);
-      const response = template.getResponse().compile();
-      callback(null, response);
-    }
-  }
-};
-
-const serverStreamController = (endpoint, mappingResolver, dataFiles) => {
-  return {
-    canHandle: async (endpointId) => {
-      return endpoint.getId() === endpointId;
-    },
-    handle : async (callContext) => {
-      const request = callContext.request;
-      const responseFile = mappingResolver.getResponseFile(endpoint.getId(), request);
-      const template = await dataFiles.get(responseFile);
-      const response = template.getResponse().compile();
-
-      if(!response){
-        return callContext.end();
-      }
-
-      const responses = [...response.stream];
-      const streamingDelay = typeof response.streamInterval === 'undefined' ? DEFAULT_STREAMING_DELAY : response.streamInterval;
-      let keepStreaming = !response.doNotRepeat ;
-
-      for(let i =0; i < responses.length || keepStreaming; i++){
-        await new Promise((resolve => setTimeout(resolve, streamingDelay)));
-        callContext.write(responses[i%responses.length]);
-      }
-
-      callContext.end();
-    }
-  }
-};
-
-const clientStreamController = (endpoint) => {
-  return {
-    canHandle: async (endpointId) => {
-
-    },
-    handle : async (endpointId, callContext) => {
-
-    }
-  }
-};
-
-const duplexStreamController = (endpoint) => {
-  return {
-    canHandle: async (endpointId) => {
-
-    },
-    handle : async (endpointId, callContext) => {
-
-    }
-  }
-};
+const UnaryController = require('./unary-controller');
+const ServerStreamController = require('./server-streaming-controller');
+const ClientStreamController = require('./duplex-streaming-controller');
+const DuplexStreamController = require('./duplex-streaming-controller');
 
 module.exports = {
   create: (endpoint, mappingResolver, dataFiles) => {
-    const type = getType(endpoint);
-
     const clientStream = endpoint.isStreamingRequest();
     const serverStream = endpoint.isStreamingResponse();
 
@@ -86,17 +13,17 @@ module.exports = {
     const isClientStream = clientStream === true && serverStream === false;
 
     if(isUnary){
-      return unaryController(endpoint, mappingResolver, dataFiles);
+      return UnaryController.create({endpoint, mappingResolver, dataFiles});
     }
 
     if(isServerStream){
-      return serverStreamController(endpoint, mappingResolver, dataFiles);
+      return ServerStreamController.create({endpoint, mappingResolver, dataFiles});
     }
 
     if(isClientStream){
-      return clientStreamController(endpoint, mappingResolver);
+      return ClientStreamController.create({endpoint, mappingResolver, dataFiles});
     }
 
-    return duplexStreamController(endpoint, mappingResolver);
+    return DuplexStreamController.create({endpoint, mappingResolver, dataFiles});
   }
 }
