+++ src/controllers/server-streaming-controller.js
@@ -1,0 +1,33 @@
+const DEFAULT_STREAMING_DELAY = 1000;
+
+module.exports = {
+  create: async ({endpoint, mappingResolver, dataFiles}) => {
+    return {
+      endpointId: endpoint.getId(),
+      canHandle: (endpointId) => {
+        return endpoint.getId() === endpointId;
+      },
+      handle : async (callContext) => {
+        const request = callContext.request;
+        const responseFile = mappingResolver.getResponseFile(endpoint.getId(), request);
+        const template = await dataFiles.get(responseFile);
+        const response = template.getResponse().compile();
+
+        if(!response){
+          return callContext.end();
+        }
+
+        const responses = [...response.stream];
+        const streamingDelay = typeof response.streamInterval === 'undefined' ? DEFAULT_STREAMING_DELAY : response.streamInterval;
+        let keepStreaming = !response.doNotRepeat ;
+
+        for(let i =0; i < responses.length || keepStreaming; i++){
+          await new Promise((resolve => setTimeout(resolve, streamingDelay)));
+          callContext.write(responses[i%responses.length]);
+        }
+
+        callContext.end();
+      }
+    }
+  }
+};
