+++ src/controllers/unary-controller.js
@@ -1,0 +1,17 @@
+module.exports = {
+  create: async ({endpoint, mappingResolver, dataFiles}) => {
+    return {
+      endpointId: endpoint.getId(),
+      canHandle: (endpointId) => {
+        return endpoint.getId() === endpointId;
+      },
+      handle : async (callContext, callback) => {
+        const request = callContext.request;
+        const responseFile = mappingResolver.getResponseFile(endpoint.getId(), request);
+        const template = await dataFiles.get(responseFile);
+        const response = template.getResponse().compile();
+        callback(null, response);
+      }
+    }
+  }
+};
